Delete order and update product stock

+++ database/migrations/2021_12_02_233235_create_category_translations_table.php
@@ -15,7 +15,7 @@ class CreateCategoryTranslationsTable extends Migration
     {
         Schema::create('category_translations', function (Blueprint $table) {
             $table->id();
-            $table->integer('category_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->string('locale')->index();
             $table->string('name');
             $table->unique(['category_id', 'locale']);
